% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adequacy.R
\name{adequacy}
\alias{adequacy}
\title{Regression predictor adequacy}
\usage{
adequacy(model = x, graph = FALSE)
}
\arguments{
\item{model}{the output from a regression model fitted using lm( ) or glm( )}

\item{graph}{logical. Default = FALSE. See return for details.}
}
\value{
A data.frame if graph = FALSE; a ggplot2 graph if graph = TRUE.
}
\description{
Regression predictor adequacy
}
\details{
Calculates the improvement in AIC that the supplied model makes compared to the NULL model (intercept only). Then fits a separate regression model for each predictor variable in the supplied model (fitted to the same cases as the full model), and for each model calculates the associated improvement in AIC over the NULL model. Finally, calculates the 'adquacy' of each predictor = AIC improvement for predictor on its own / AIC improvement of full model x 100. Based on Harrell F E (2015) Regression modeling strategies, 2nd ed., New York: Springer, pp. 207-8.
}
\examples{
# The adequacy of predictors in an OLS regression model
lm( Sex ~ Age + Tenure + Income, data = survey ) |>
 adequacy( )

# The adequacy of predictors in a logistic regression model
glm( Sex ~ Age + Tenure + Income, data = survey,
     family = stats::binomial( link = "logit" ) ) |>
  adequacy( )

}
