% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_fit_line.R
\name{best_fit_line}
\alias{best_fit_line}
\title{Tool to explore how best-fit line works in OLS regression}
\usage{
best_fit_line(df = NULL, a, b)
}
\arguments{
\item{df}{an optional data.frame (default = NULL). If supplied, first column treated as X variable and second column as Y variable.}

\item{a}{numeric. Height of intercept on y-axis when x = 0.}

\item{b}{numeric. Slope of line.}
}
\value{
a ggplot2 graph showing fit of line with intercept a and slope b to data plotted in graph. Also reports formula for line plus Total Sum of Squares to console.
}
\description{
Tool to explore how best-fit line works in OLS regression
}
\examples{
best_fit_line( a = 10, b = 0 )

}
