% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffs.R
\name{coeffs}
\alias{coeffs}
\title{Convert regression model coefficients and fits into a data.frame}
\usage{
coeffs(..., measure = "all", fit = NULL)
}
\arguments{
\item{...}{one or more lm or glm regression models}

\item{measure}{character vector. One or more of the measures made available from broom::tidy( model ). Default value of "all" ensures all available measure are returned.}

\item{fit}{character vector. One or more of the measures of model fit made available from broom::glance( model). Default value NULL returns none. Value of "all" returns all available measures of fit.}
}
\value{
a data.frame containing the requested measures and fits for each supplied regression model.
}
\description{
Convert regression model coefficients and fits into a data.frame
}
\examples{
model1 <- lm( Height ~ Income, data = survey )
model2 <- lm( Height ~ Income + Sex, data = survey )
model3 <- glm( Sex ~ Age,
               data = survey,
               family = binomial( link = "logit" ) )

# Capture all measures for both models, but no fits
coeffs( model1, model2 )

# Capture selected measures and fits
coeffs( model1, model2,
        measure = c("estimate", "p.value"),
        fit = c("r.squared", "adj.r.squared", "AIC") )

# Capture regression coefficients only
coeffs( model1, model2, measure = "estimate" )

# Capture all measures and fits
coeffs( model1, model2, measure = "all", fit = "all" )

# Additional measure (expB) and fit (tjur.rsq) available for
# logistic regression models
coeffs( model3, measure = "all", fit = "all" )

# Extract odds ratio and AIC only for logistic regression model
coeffs( model3, measure = "expB", fit = "AIC" )

}
