% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffs_pub.R
\name{coeffs_pub}
\alias{coeffs_pub}
\title{Publication-standard table of regression coefficients}
\usage{
coeffs_pub(df, line_width = 1.2)
}
\arguments{
\item{df}{a data.frame containing regression model results produced by the gdslStats::coeff function.}

\item{line_width}{numeric. Thickness of horizontal lines in table.}
}
\value{
a flextable object.
}
\description{
Publication-standard table of regression coefficients
}
\examples{
model <- lm( Income ~ Height, data = survey )
coeffs( model ) |> coeffs_pub( )

}
