% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_df.R
\name{convert_df}
\alias{convert_df}
\title{Convert a data.frame from matrix to long format and retain results only for specified variable(s)}
\usage{
convert_df(df, statistic, var_name = NULL)
}
\arguments{
\item{df}{a data.frame containing the results from a correlation matrix or similar}

\item{statistic}{a character string naming the column to be used for storing values.}

\item{var_name}{a character string giving the name of a variable in df for which all pair-wise results are required. Default is NULL, which results in the return of all possible pairwise results.}
}
\value{
a 'long' data.frame containing the columns x, y and statistic.
}
\description{
Convert a data.frame from matrix to long format and retain results only for specified variable(s)
}
\examples{
# Create a matrix-style data.frame of correlation results
#df <- survey |>
#  dplyr::select_if( is.numeric ) |>
#  cor( "pearson" )
#df

# Convert to long-format
#df |> convert_df( statistic = "r" )

# Convert to long-format, retaining only results pertaining to var_name
#df |> convert_df( statistic = "r", var_name = "Height" )

}
