% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor}
\alias{cor}
\title{Correlation (Pearson, Spearman or Cramer's V)}
\usage{
cor(
  y,
  method = NULL,
  use = "everything",
  useNA = "no",
  p_value = FALSE,
  sig = FALSE,
  ci = NULL,
  var_name = NULL,
  sort_by = "r",
  sort_abs = FALSE,
  matrix = FALSE,
  fill = TRUE
)
}
\arguments{
\item{y}{a data.frame containing two or more columns}

\item{method}{a character string indicating the type of correlation to be calculated. One of "pearson", "spearman" or "cramer".}

\item{use}{a character string giving a method for computing correlations in the presence of missing values. This must be one of "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs". Default is "everything". See Details.}

\item{useNA}{a character string, one of "no" and "ifany" indicating whether NA should be treated as a category when calculating Cramer's V. Default is "no".}

\item{p_value}{logical (TRUE or FALSE), indicating whether results should include p-values. Default is FALSE.}

\item{sig}{logical (TRUE or FALSE), indicating whether results should include asterisks flagging 'significant' values (i.e. those with p-values less than 0.01, 0.05 and 0.1). Defaut is FALSE.}

\item{ci}{NULL (default) or a numeric value in the range 0 to 100 indicating what \% confidence intervals are required. (e.g 95 for 95\% CIs)}

\item{var_name}{a character string giving the name of a variable in y for which all pair-wise correlations are required. Default is NULL, which results in the return of all possible pairwise correlations.}

\item{sort_by}{a character string indicating on which column the output should be sorted by. Options are "x", "y", "r", "p", "s", "ci.lb" and "ci.ub". Default is "r".}

\item{sort_abs}{logical (TRUE or FALSE). Default is FALSE. If TRUE, output sorted on a column containing numeric values will be sorted on the absolute value of thise numeric values (i.e. ignoring the sign (+/-) of the values).}

\item{matrix}{logical (TRUE or FALSE). Default is FALSE. If TRUE, instead of output being returned in one data.frame, the results are returned as a list of data.frames (one per type of output requested), in a format suitable for direct input to cor_plot( ).}

\item{fill}{logical (TRUE or FALSE). Default is TRUE. If FALSE, only the top half of each matrix is completed. Applies only if matrix = TRUE.}
}
\value{
If matrix = FALSE, one data.frame containing all results in 'long' format. If y contains only two columns, then only one row of results returned (col1 x col2), instead of all possible correlations (col1 x col1, col1 x col2, col2 x col1, col2 x col2). If matrix = TRUE, a list of data.frames is returned, one per requested measure, and the data.frame will be in 'matrix' format.
}
\description{
Correlation (Pearson, Spearman or Cramer's V)
}
\examples{
# Correlation of two variables
survey |> dplyr::select( Income, Height ) |> cor( "pearson" )

# Correlation between all variables in a supplied data.frame,
# plus return of p-value, sig and CIs in addition to r.
#survey |>
#   dplyr::select_if( is.numeric) |>
#   cor( "spearman", p_value = TRUE, sig = TRUE, ci = 95 )

# As above, but with results filtered to report only the correlations
# with one named variable, and sorted in order of the contents of column y.
#survey |>
#   dplyr::select_if( is.numeric) |>
#   cor( "spearman", p_value = TRUE, sig = TRUE, ci = 95,
#        var_name = "Height", sort_by = "y" )

# Results reported in matrix format (suitable as an input to cor_plot)
# survey |>
#   dplyr::select_if( is.numeric ) |>
#   cor( "spearman", p_value = TRUE, sig = TRUE, ci = 95,
#        matrix = TRUE )

}
