% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{Plot a correlogram}
\usage{
cor_plot(
  x,
  style = "circle",
  sig = FALSE,
  sig_col = "black",
  sig_size = 2,
  label_size = 1
)
}
\arguments{
\item{x}{a list supplying a set of correlations and p-values, each in a 'matrix' style data.frame, as produced by the gdsdlStats::cor function when matrix = TRUE and p_value = TRUE.}

\item{style}{a character string specifying correlogram style: one of "circle" (the default) and "heatmap".}

\item{sig}{logical (TRUE/FALSE) indicating whether statistically significant correlations should be flagged using asterisks (for the p<0.01, p<0.05 and p<0.1 levels). Default is FALSE.}

\item{sig_col}{a character string specifying the colour of any asterisks. Default is "black". For other options run grDevices::colors().}

\item{sig_size}{a numeric value (default 2) indicating the size of any asterisks.}

\item{label_size}{a numeric value (default 1) indicating the size of the text used to label variables associated with each row and column.}
}
\value{
a correlogram in corrplot format
}
\description{
Plot a correlogram
}
\details{
A wrapper for the \code{corrplot( )} function from the \code{corrplot} package.
}
\examples{
# Default cor_plot
survey |>
  dplyr::select_if( is.numeric ) |>
  cor( "pearson", p_value = TRUE, matrix = TRUE) |>
  cor_plot( )

# cor_plot in  heatmap format with significant correlations flagged
# and changes to text and asterisk size
survey |>
  dplyr::select_if( is.numeric ) |>
  cor( "pearson", p_value = TRUE, matrix = TRUE) |>
  cor_plot( style = "heatmap", sig = TRUE )

}
