% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test.R
\name{cor_test}
\alias{cor_test}
\title{Calculate a pair-wise correlation and its associated p-value, significance and Confidence Interval}
\usage{
cor_test(x, y, method, p_value = FALSE, sig = FALSE, ci = NULL, useNA = "no")
}
\arguments{
\item{x}{a numeric, character or factor vector (one of the two sets of values being correlated)}

\item{y}{a numeric, character or factor vector (one of the two sets of values being correlated)}

\item{method}{a character string indicating the type of correlation to be calculated. One of "pearson", "spearman" or "cramer".}

\item{p_value}{logical (TRUE/FALSE), indicating whether or not p-value should be calculated. Default is FALSE.}

\item{sig}{logical (TRUE/FALSE), indicating whether or not the sigificance of p-value should be calculated. Default is FALSE.}

\item{ci}{default is NULL. Otherwise a numeric value between 0 and 100 indicating the type of Confidence Interval required (e.g. 95\% CI).}

\item{useNA}{a character string, one of "no" and "ifany" indicating whether NA should be treated as a category when calculating Cramer's V. Default is "no".}
}
\value{
a one-row data.frame containing the pairwise correlation value, plus p_value, sig and ci if requested.
}
\description{
Calculate a pair-wise correlation and its associated p-value, significance and Confidence Interval
}
\examples{
# A simple pairwise correlation
cor_test( survey$Income, survey$Height, method = "pearson" )

# Same correlation with added information
cor_test( survey$Income, survey$Height, method = "pearson",
          p_value = TRUE, sig = TRUE, ci = 95 )

# Correlation of two categorical variables, ignoring any NA values
df <- survey
df[ 1:100 , "Sex"] <- NA
cor_test( df$Age, df$Sex, method = "cramer" )

# Correlation of two categorical variables, treating NA as a valid category
cor_test( df$Age, df$Sex, method = "cramer", useNA = "ifany" )

}
