% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramersV_test.R
\name{cramersV_test}
\alias{cramersV_test}
\title{Calculate Cramer's V and associated p-value for a variable pair}
\usage{
cramersV_test(
  x,
  y,
  p_value = TRUE,
  sig = FALSE,
  ci = NULL,
  useNA = "no",
  NaN_warning = TRUE
)
}
\arguments{
\item{x}{a factor, character, logical or numeric vector (one of variables being correlated).}

\item{y}{a factor, character, logical or numeric vector (one of variables being correlated).}

\item{p_value}{logical (TRUE or FALSE), indicating whether results should include p-values. Default is FALSE.}

\item{sig}{logical (TRUE or FALSE), indicating whether results should include asterisks flagging 'significant' values (i.e. those with p-values less than 0.01, 0.05 and 0.1). Defaut is FALSE.}

\item{ci}{NULL (default) or a numeric value in the range 0 to 100 indicating what \% confidence intervals are required. (e.g 95 for 95\% CIs) NOTE: THIS FEATURE NOT YET ENABLED.}

\item{useNA}{a character string, one of "no" and "ifany" indicating whether NA should be treated as a category when calculating Cramer's V. Default is "no".}

\item{NaN_warning}{logical (default TRUE), to give warning if a Cramer's V value of NaN arises from an empty row/column in the cross-tabulation of x by y.}
}
\value{
a one-row data.frame containing the correlation plus any requested associated measures.
}
\description{
Calculate Cramer's V and associated p-value for a variable pair
}
\examples{
cramersV_test( survey$Age, survey$Sex, p_value = TRUE, sig = TRUE )

}
