% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_values.R
\name{expected_values}
\alias{expected_values}
\title{Title Report the (minimum) expected values under-pinning a glm( ) regression model}
\usage{
expected_values(model)
}
\arguments{
\item{model}{The output from a model fitted using the glm( ) function}
}
\value{
Writes results to screen; invisibly returns in list format the following three values: min_expected (the minimum expected value); pct_5plus (the \% of cells containing a count of 5 or more); expected_values (a data.frame reporting the observed and expected cell count for each cell in a cross-tabulation of the model outcome variable with all of the model predictor variables )
}
\description{
Title Report the (minimum) expected values under-pinning a glm( ) regression model
}
\examples{
# Fit glm model
model <-
  stats::glm( Sex ~ Tenure, data = survey,
              family = stats::binomial( link = "logit" ),
              na.action = stats::na.omit )

## Identify minimum expected value and \% of expected counts < 5
expected_values( model )

## Inspect the invisibly returned set of results
res <- expected_values( model )
res

}
