% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.categorical.R
\name{is.categorical}
\alias{is.categorical}
\title{Check whether vector is categorical}
\usage{
is.categorical(x, num = 20)
}
\arguments{
\item{x}{a vector}

\item{num}{numeric (default = 20). Numeric vectors with fewer than num unique values will be classed as categorical.}
}
\value{
TRUE if vector is factor, character, logical, or a numeric vector with less than num unique values; else FALSE.
}
\description{
Check whether vector is categorical
}
\examples{
# Applied to a factor vector
is.categorical( survey$Age )

# Applied to a logical vector
is.categorical( c( TRUE, FALSE, TRUE, TRUE, FALSE ) )

# Applied to a character vector
is.categorical( c( "A", "B", "C", "A", "B", "C" ) )

# Applied to a numeric vector with less than num = 20 unique values
is.categorical( c( 1, 2, 3, 4, 5 ) )

# Applied to a numeric vector with less than num = 0 unique values
is.categorical( c( 1, 2, 3, 4, 5 ), num = 0 )

# Used to select the categorical variables in a data.frame
survey |> dplyr::select_if( is.categorical ) |> names( )

}
