% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.symmetrical.R
\name{is.symmetrical}
\alias{is.symmetrical}
\title{Check whether a vector is skewed ( abs(skew) < threshold )}
\usage{
is.symmetrical(x, threshold = 0.5)
}
\arguments{
\item{x}{a numeric vector}

\item{threshold}{a numeric value (default 0.5).}
}
\value{
logical (TRUE or FALSE). TRUE if absolute value of skew is less than or equal to the threshold; else FALSE.
}
\description{
Check whether a vector is skewed ( abs(skew) < threshold )
}
\examples{
# Check to see whether a given variable is symmetrical
is.symmetrical( survey$Height )

#Use is.symmetrical to select only the skewed variables in a given data.frame
survey |> dplyr::select_if( is.symmetrical ) |> names( )

}
