% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.unordered.R
\name{is.unordered}
\alias{is.unordered}
\title{Check to see if vector is unordered}
\usage{
is.unordered(x)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE if vector is logical, character, or a factor that is not ordered; else FALSE.
}
\description{
Check to see if vector is unordered
}
\examples{
is.unordered( survey$Sex )
is.unordered( survey$Age )
survey |> dplyr::select_if( is.unordered ) |> names( )

}
