% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.R
\name{median}
\alias{median}
\title{Median value of a numeric or ordinal variable}
\usage{
median(x, na.rm = FALSE)
}
\arguments{
\item{x}{A \emph{numeric}, \emph{factor} or \emph{character} vector.}

\item{na.rm}{If \code{FALSE} (default) observations with a missing value of \code{x} are NOT removed before applying the test. If \code{TRUE}, missing values of \code{x} ARE removed before applying test.}
}
\value{
Returns the calculated median value (\emph{Numeric}) or category (\emph{character}).
}
\description{
Calculates the median of a set of \emph{numeric} or \emph{categorical} values, first converting any \emph{character} variable into a factor
}
\note{
A wrapper for the \code{\link{stats::median}} function supplied as part of the base R \code{stats} package, adding the ability to handle categorical (factor) vectors.
}
\examples{
# Median of numeric vector
median( survey$Income )

# Median of an ordinal vector (ordered factor)
median( survey$Age )

# Median of a nominal vector (unordered factor) which includes missing values
median( survey$WorkStatus, na.rm = TRUE )
}
\author{
Paul Williamson
}
