% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_plots.R
\name{outcome_plots}
\alias{outcome_plots}
\title{Scatterplot matrix with specified variable on y-axis}
\usage{
outcome_plots(df, outcome_name)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{outcome_name}{Character. The name of the variable to be plotted on the y-axis, in quotes.}
}
\value{
A list containing a series of scatterplots, one fewer than the number of variables in the supplied data.frame, with each plot having the variable identified by outcome_name on the y-axis and one of the other variables in the supplied data.frame on the x-axis. For scatterplots involving numeric x and y variables, the scatterplot includes an added linear line of best-ft.
}
\description{
Scatterplot matrix with specified variable on y-axis
}
\examples{
survey |>
  outcome_plots( "Height" ) |>
  patchwork::wrap_plots( )

}
