% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_tab.R
\name{publish_tab}
\alias{publish_tab}
\title{Convert \code{tab} object to publication standard}
\usage{
publish_tab(res, line_width = 1.2)
}
\arguments{
\item{res}{A tabulation stored as a data object of class \code{tab} (e.g. the output from the \code{tab} function).}

\item{line_width}{The line width to use when drawing any horizontal lines in the table (Default is 1.2).}
}
\value{
A publication-standard version of the input table, returned as a \code{flextable} data object.
}
\description{
Convert \code{tab} object to publication standard
}
\examples{
# Create a tabulation using the `tab` function
example_table <- survey |>
  tab( Age ~ Sex, measure = "row_pct", totals = "row", base_count = TRUE )

# Create publication standard version
publish_tab( example_table )

# As above, but with rounding (which must happen before publish_tab is called)
example_table |> round( dp = c( 1, 1, 0, 0 ) ) |> publish_tab( )
}
