% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{OLS and Logistic Regression}
\usage{
regress(x, f, method = "ols")
}
\arguments{
\item{x}{a data.frame or tibble}

\item{f}{a regression formula}

\item{method}{character. Regression method - one of "ols" or "logistic". Default is "ols".}
}
\value{
The output from an OLS regression model fitted using lm( ), of from a logistic regression model fitted using glm( family=binomial(link = "logit" ) )
}
\description{
OLS and Logistic Regression
}
\examples{
# Convert Sex from a two-category factor variable to a logical variable,
# so that it can be used as outcome for OLS regression
survey$Female <- dplyr::if_else( survey$Sex == "Female", TRUE, FALSE )

# Fit default regression model (OLS)
survey |> regress( Female ~ Height )

# Fit a logistic regression model
survey |> regress( Female ~ Height, "logistic" )

}
