% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_and_CIs.R
\name{sample_size_and_CIs}
\alias{sample_size_and_CIs}
\title{Visualisation of the impact of sample size on the 95\% CIs of percentages}
\usage{
sample_size_and_CIs(pct_name, p, y_min = NULL, y_max = NULL)
}
\arguments{
\item{pct_name}{Character. Name of percentage - eg. Unemployed.}

\item{p}{Numeric. A percentage - e.g. 34}

\item{y_min}{Numeric. Minimum y-axis value. Default value of \code{NULL} results in \code{y_min} being set to the lowest confidence interval value, rounded down to the nearest integer.}

\item{y_max}{Numeric. Maximum y-axis value. Default value of \code{NULL} results in \code{y_max} being set to the highest confidence interval value, rounded up to the nearest integer.}
}
\value{
A ggplot graph visualisng the 95\% Confidence Interval for the supplied value of \code{p} for a set of sample sizes ranging from 250 to 10,000.
}
\description{
Visualisation of the impact of sample size on the 95\% CIs of percentages
}
\examples{
# Default y-axis limits
sample_size_and_CIs( "Unemployed", 34 )

# User-defined y-axis limits
sample_size_and_CIs( "Unemployed", 34, 0, 100 )
}
