% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_distribution.R
\name{sampling_distribution}
\alias{sampling_distribution}
\title{Simulation of sampling distribution and Confidence Intervals for percentages}
\usage{
sampling_distribution(x, category, no_of_samples, sample_size)
}
\arguments{
\item{x}{A character or factor vector or single-column data-frame or tibble.}

\item{category}{Character. Name of one of the categories/levels in \emph{x}.}

\item{no_of_samples}{Numeric. The number of samples to be taken from \emph{x}.}

\item{sample_size}{Numeric. The size of sample to be taken from \emph{x}.}
}
\value{
A list comprising the following set of objects:
\itemize{
\item \code{samples}  a data-frame storing the sample-specific results
\item \code{distribution_stats} sampling distribution summary statistics
\item \code{distribution} a histogram of the sampling distribution, including the \% of \code{category} observed in \code{x} and mean of the sample-based estimates of this quantity
\item \code{coverage_graph} a plot illustrating the calculated CIs, and flagging in red those that do NOT cross include the observed value in \code{x}
\item \code{coverage_rate} the 95\% CI coverage rate (\% of 95\% CIs that include th observed value in \code{x})
}
}
\description{
Simulation of sampling distribution and Confidence Intervals for percentages
}
\details{
Calculates the percentage of the values of \code{x} that fall in the specified \code{category}.
Then takes \code{no_of_samples} of \code{ sample_size} from the supplied vector \emph{x}. For each sample,
calculates the percentage of the sample values which fall in the specified \code{category}, and calculates
the associated standard error (of proportions) and 95\% confidence interval. Finally, summarises the
results in a series of plots and summary statistics. The purpose of the function is to allow the user
to explore how \code{no_of_samples} and \code{sample_size} interact to influence the accuracy of the
sampling distribution mean, plus the sample-specific calculated confidence intervals,
and the overall 95\% CI coverage rates.
}
\examples{
# Generate a vector comprising values of "M" and "F"
x <- sample( c("M", "F"), 1000, replace = TRUE )

# Find the sampling distribution for the percentage of values that fall in
# category "F", using 50 random samples of size 20
output <- sampling_distribution( x, "F", no_of_samples = 50, sample_size = 20 )

# First few rows of the samples data.frame
output$samples |> head( )

# Summary stats for sampling distribution
output$distribution_stats

# Histogram of sampling distribution
output$distribution

# Coverage graph
output$coverage_graph

# Coverage rate
output$coverage_rate

}
