% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd}
\alias{sd}
\title{Sample or Population Standard Deviation}
\usage{
sd(x, na.rm = FALSE, sample = TRUE)
}
\arguments{
\item{x}{A \emph{numeric} vector.}

\item{na.rm}{If \code{FALSE} (default) observations with a missing value of \code{x} are NOT removed before applying the test. If \code{TRUE}, missing values of \code{x} ARE removed before applying test.}

\item{sample}{If \code{TRUE} (default), calculation uses \emph{n}-1 in the denominator. If \code{FALSE}, uses \emph{n} instead, as per standard statistical theory.}
}
\value{
\emph{Numeric}. Returns the value of the calculated standard deviation.
}
\description{
Calculates the standard deviation of a set of \emph{numeric} values, making an appropriate adjustment if using population rather than sample data. (The value returned by \link[stats:sd]{stats::sd} only supports division by \emph{n}-1)
}
\examples{
# Calculate standard deviation of a sample (default)
sd( survey$Height )

# For a sample, std.dev returns same result as stats::sd function
sd( survey$Height ) == stats::sd( survey$Height )

# Calculate standard deviation of a population
sd( survey$Height, sample = FALSE )

}
