% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{skew}
\title{Skew}
\usage{
skew(x, na.rm = FALSE)
}
\arguments{
\item{x}{A \emph{numeric} vector.}

\item{na.rm}{If \code{TRUE} observations with a missing value of \code{x} are removed before applying the test. Default is \code{FALSE}.}
}
\value{
\emph{Numeric}. Returns the value of the calculated skew.
}
\description{
Calculates the skew of a set of \emph{numeric} values (assumed to be from a sample).
}
\note{
A near direct copy of solution posted by Wolfgang Koller at \url{https://stat.ethz.ch/pipermail/r-help/1999-July/004529.html}.
}
\examples{
# Calculate the skew of a sample
skew( survey$Height )

}
