% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_df.R
\name{sort_df}
\alias{sort_df}
\title{Sort data.frame containing results from the cor function}
\usage{
sort_df(df, sort_by = "r", sort_abs = FALSE)
}
\arguments{
\item{df}{a data.frame containing results from the cor function.}

\item{sort_by}{a character string indicating on which column the output should be sorted by. Options are "x", "y", "r", "p", "s", "ci.lb" and "ci.ub". Default is "r".}

\item{sort_abs}{logical (TRUE or FALSE). Default is FALSE. If TRUE, output sorted on a column containing numeric values will be sorted on the absolute value of thise numeric values (i.e. ignoring the sign (+/-) of the values).}
}
\value{
a sorted data.frame
}
\description{
Sort data.frame containing results from the cor function
}
\examples{
# Create data.frame capturing outputs from gdslStats::cor function
#df <- survey |>
#  dplyr::select_if( is.numeric ) |>
#  cor( "pearson", p_value = TRUE, sig = TRUE, ci = 95 )

 # Use sort_df to sort by p_value
#  df |> sort_df( sort_by = "p" )

}
