% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srdi.R
\name{srdi}
\alias{srdi}
\title{Standardised Reciprocal Diversity Index}
\usage{
srdi(x, na.rm = TRUE)
}
\arguments{
\item{x}{A \emph{character} or \emph{factor} vector. If one or more categories lack observations, see Details.}

\item{na.rm}{If \code{FALSE} (default) observations with a missing value of \code{x} are NOT removed. If \code{TRUE}, missing values of \code{x} ARE removed.}
}
\value{
\emph{Numeric}. Returns the value of the calculated Standardised Reciprocal Diversity Index
}
\description{
Standardised Reciprocal Diversity Index
}
\details{
Calculates the value of the SRDI assuming that there are no categories without observations. If there are, supply the data as a factor which includes the empty categories as named factor levels.
}
\examples{
# Calculate SRDI when values evenly spread across categories
x <- factor( c("A", "A", "B", "B", "C", "C"), levels = LETTERS[1:3] )
srdi( x )

# Calculate SRDI when values are unevenly spread across categories
x <- factor( c("A", "B", "C", "C", "C", "C"), levels = LETTERS[1:3] )
srdi( x )

# Calculate SRDI when all values are maximally concentrated
# (Note the need to declare the empty categories as factor levels)
x <- factor( c("A", "A", "A", "A", "A", "A"), levels = LETTERS[1:3] )
srdi( x )

}
