% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{survey}
\alias{survey}
\title{Simulated Survey Data}
\format{
A data frame containing 1000 persons with 11 variables.
}
\description{
This is a fictional example of a social survey dataset, provided for demonstration purposes. It consists of variables for unique Person and Houeshold IDs, Tenure, Age, Sex, Health, WorkStatus, Height, and Income, followed by household- and person-level initial weights (Household_w0 and Person_w0).
}
\keyword{datasets}
