% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{(Cross-)tabulate survey and admin data}
\usage{
tab(
  data = NULL,
  formula = NULL,
  weights,
  measure = "count",
  totals = "none",
  cum_sum = FALSE,
  base_count = FALSE,
  total_3d = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{data}{A data.frame or tibble containing the dataset to be tabulated}

\item{formula}{A formula specifying the tabulation required, of the form: ~ x, or y ~ x, or y ~ x + z}

\item{weights}{The name of the \code{data} variable, if any, to be used to weight the tabulation.}

\item{measure}{The type of tabulation required. Specifies both the metric to be used (count, percentage or proportion) and the directionality of this metric (row, column or joint). Options are "count", "row_pct", "col_pct", "joint_pct", "row_prop", "col_prop" and "joint_prop".}

\item{totals}{Which totals (table margins), the table should include. Options are: "none" (the default), row", "col" and "row_col".}

\item{cum_sum}{Whether or not values should be cumulated across the cells of the table. Options are FALSE (the default) or TRUE.}

\item{base_count}{Whether or not the base count associated with any percentages or proportions should be included in the table. Options are FALSE (the default) or TRUE.}

\item{total_3d}{Whether or not three-way tables should include a final two-way table summing values across the third table dimension. Options are FALSE (the default) or TRUE.}

\item{na.rm}{Whether or not NA (missing) values should be excluded from the table. Options are FALSE (the default) or TRUE.}
}
\value{
A one-, two-, or three-way tabulation, supplied as an object of class table, xtable and tab.
}
\description{
(Cross-)tabulate survey and admin data
}
\examples{
# Load toy dataset supplied with package
#load(survey)

# Tabulation examples

# Frequency table (counts), with and without missing values
survey |> tab( ~WorkStatus, na.rm = TRUE )
survey |> tab( ~WorkStatus, na.rm = FALSE )

# Freuquency table (proportions and cumulative proportions)
survey |> tab( ~ Age, measure = "col_prop" )
survey |> tab( ~ Age, measure = "col_prop", cum_sum = TRUE )

# Two-way cross-tabulation (column percentages; row percentages with row totals)
survey |> tab( Sex ~ Age, measure = "col_pct" )
survey |> tab( Sex ~ Age, measure = "row_pct", totals = "row_col" )

# Two-way cross-tabulation (row percentages with row totals and base count)
survey |> tab( Sex ~ Age, measure = "row_pct", totals = "col",
               base_count = TRUE )

# Three-way cross-tabulation including counts aggregated across third dimension
survey |> tab( Health ~ Age + Sex, measure = "count", total_3d = TRUE )

# All-singing all-dancing three-way table
survey |> tab( Health ~ Age + Sex, measure = "row_pct", totals = "row",
               base_count = TRUE, total_3d = TRUE )
}
