% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjurs_rsq.R
\name{tjurs_rsq}
\alias{tjurs_rsq}
\title{Tjur's R-squared - a measure of fit for logistic regression models}
\usage{
tjurs_rsq(x)
}
\arguments{
\item{x}{a glm logit model}
}
\value{
numeric - Tjur's r squared, expressed as a percentage
}
\description{
Tjur's R-squared - a measure of fit for logistic regression models
}
\examples{
# Fit glm model
model <- glm( Sex ~ Age, data = survey,
              family = binomial( link = "logit" ) )
tjurs_rsq(  model )

}
