% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_gleman.R
\name{z_gelman}
\alias{z_gelman}
\title{Gelman standardisation}
\usage{
z_gelman(x)
}
\arguments{
\item{x}{A numeric vector containing at least two elements}
}
\value{
A numeric vector containing Gelman standardised values of x.
}
\description{
Gelman standardisation
}
\details{
A function to standardise a numeric variable to have a mean of 0 and an sd of 0.5, by subtracting the mean of variable and then dividing by 2 x standard deviation, ignoring any NA values. Based on the standardisation approach suggested by Gelman A and Hill J (2007) Data analysis using regression and multilevel/hierarchical models, Cambridge:Cambridge University Press, pp. 54-57.
}
\examples{
z_gelman( c(1:3, NA, 5:7, NA, 9:10) )

#' # Create Gelman standardisd version of Height
survey |>
  dplyr::mutate( z_Height = z_gelman( Height ) ) |>
  dplyr::select( Height, z_Height ) |>
  dplyr::slice( 1:10 )

}
