% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round}
\alias{round}
\title{Round numerical data}
\usage{
round(data_in, dp = 2, as_string = TRUE)
}
\arguments{
\item{data_in}{The data to be rounded, supplied as a (vector of) numeric or logical; or as data object of class table, matrix, data.frame or tibble (any of which can contain non-numeric columns)}

\item{dp}{The number of decimal places to round to (default is 2), supplied as either as single number to be applied to all numeric values, or as a series vector of values (one per column of the input data object, in column order, using NA for factor/string columns)}

\item{as_string}{If TRUE (default is FALSE), and data_in is of class table, returns a table in which all numeric values have been converted to strings, so that the dp per column used for display matches that used for rounding}
}
\value{
An updated version of the supplied data object containing rounded numeric values
}
\description{
Round numerical data
}
\examples{
round( data_in = 1.1234 )
round( data_in = rep(1.1234, 4), dp = c(1,2,3,4) )
df <- data.frame( var = c("A", "B", "C"), v1 = rep(1.1234, 3), v2 = rep(1.1234, 3) )
round( df, dp = c(2, 1) )
# Example of table with as_string = FALSE to follow
round( tibble::as_tibble( df), dp = c(NA, 2, 1) )

}
